requirements.txt
py-clob-client
web3
requests
websocket-client